+++ ui/main/game/galactic_war/cards/gwc_start_subcdr.js
@@ -28,17 +28,15 @@ define([
         faction: inventory.getTag("global", "playerFaction") || 0,
       };
     },
-    deal: function (__, context) {
-      var minions = _.shuffle(GWFactions[context.faction].minions.slice(0));
+    deal: function () {
       return {
         params: {
-          minions: minions.slice(0, 2),
           allowOverflow: true,
         },
         chance: 0,
       };
     },
-    buff: function (inventory, context) {
+    buff: function (inventory) {
       if (inventory.lookupCard(CARD) === 0) {
         // Make sure we only do the start buff/dull once
         var buffCount = inventory.getTag("", "buffCount", 0);
@@ -46,17 +44,10 @@ define([
           GWCStart.buff(inventory);
           if (inventory.getTag("global", "playerFaction") === 4)
             gwaioFactionCluster.buff(inventory);
-          inventory.maxCards(inventory.maxCards() - 2);
           inventory.addUnits([
             "/pa/units/land/vehicle_factory/vehicle_factory.json",
             "/pa/units/land/tank_light_laser/tank_light_laser.json",
           ]);
-          // eslint-disable-next-line lodash/prefer-map
-          _.forEach(context.minions, function (minion) {
-            inventory.minions.push(minion);
-          });
-          var minionSpecs = _.compact(_.pluck(context.minions, "commander"));
-          inventory.addUnits(minionSpecs);
         }
         ++buffCount;
         inventory.setTag("", "buffCount", buffCount);